RF: Move to base module, isolate error message
